// src/app/products/products.interface.ts
export interface ProductInterface {
    id: number;
    price: number;
    productName: string;
    brand: string;
    brandName: string;
    imagePath: string;
    description: string;
    scent: string;
    type: any[];
    colorType: any[];
    modalImagePath: string;
    discount10: boolean;
}